+++ src/assets/stylesheets/media-browser.scss
@@ -58,7 +58,7 @@
 
       i {
         color: $duck-yellow;
-        margin-right: 8px;
+        margin-right: 12px;
       }
     }
 
@@ -496,4 +496,14 @@
       height: 35px;
     }
   }
+}
+
+:local(.media-browser-inline) {
+  position: relative;
+  background-color: rgba(255, 255, 255, 0.6);
+  overflow-x: auto;
+  overflow-y: auto;
+  padding: 12px;
+  padding-top: 20px;
+  margin-bottom: 52px;
 }
