+++ src/index.js
@@ -5,7 +5,7 @@ import "./assets/stylesheets/index.scss";
 import registerTelemetry from "./telemetry";
 import HomeRoot from "./react-components/home-root";
 import AuthChannel from "./utils/auth-channel";
-import { createAndRedirectToNewHub, connectToReticulum } from "./utils/phoenix-utils";
+import { createAndRedirectToNewHub, connectToReticulum, fetchReticulum } from "./utils/phoenix-utils";
 import Store from "./storage/store";
 
 const qs = new URLSearchParams(location.search);
@@ -15,9 +15,13 @@ const { pathname } = document.location;
 const sceneId = qs.get("scene_id") || (pathname.startsWith("/scenes/") && pathname.substring(1).split("/")[1]);
 
 const store = new Store();
+window.APP = { store };
+
 const authChannel = new AuthChannel(store);
 let installEvent = null;
+let favoriteHubsResult = null;
 let mountedUI = false;
+let hideHero = true;
 
 const remountUI = function() {
   mountedUI = true;
@@ -33,6 +37,8 @@ const remountUI = function() {
       authToken={qs.get("auth_token")}
       authOrigin={qs.get("auth_origin")}
       listSignup={qs.has("list_signup")}
+      hideHero={hideHero}
+      favoriteHubsResult={favoriteHubsResult}
       report={qs.has("report")}
       installEvent={installEvent}
     />
@@ -57,5 +63,14 @@ window.addEventListener("beforeinstallprompt", e => {
   }
 
   authChannel.setSocket(await connectToReticulum());
+  remountUI();
+
+  if (authChannel.signedIn) {
+    // Fetch favorite rooms
+    const path = `/api/v1/media/search?source=favorites&type=hubs&user=${store.credentialsAccountId}`;
+    favoriteHubsResult = await fetchReticulum(path);
+  }
+
+  hideHero = false;
   remountUI();
 })();
